requirements.txt
coverage==7.6.9
iniconfig==2.0.0
MouseInfo==0.1.3
packaging==24.2
pillow==11.0.0
pluggy==1.5.0
PyAutoGUI==0.9.54
PyGetWindow==0.0.9
PyMsgBox==1.0.9
pyobjc-core==10.3.2
pyobjc-framework-Cocoa==10.3.2
pyobjc-framework-Quartz==10.3.2
pyperclip==1.9.0
PyRect==0.2.0
PyScreeze==1.0.1
pytesseract==0.3.13
pytest==8.3.4
pytweening==1.2.0
PyYAML==6.0.2
rubicon-objc==0.4.9
six==1.17.0
